refactor: make reusable some parts of the vesting tests
