+++ app/test/vesting_test.go
@@ -22,7 +22,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 	tmtime "github.com/tendermint/tendermint/types/time"
-	"google.golang.org/grpc"
 )
 
 const (
@@ -103,7 +102,7 @@ func (s *VestingModuleTestSuite) TestGenesisDelayedVestingAccountsSpendableBalan
 		name := s.unusedAccount(DelayedVestingAccountType)
 		address := getAddress(name, s.cctx.Keyring).String()
 
-		resAccBytes, err := GetRawAccountInfo(s.cctx.GRPCClient, address)
+		resAccBytes, err := testfactory.GetRawAccountInfo(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		var vAcc vestingtypes.DelayedVestingAccount
@@ -112,7 +111,7 @@ func (s *VestingModuleTestSuite) TestGenesisDelayedVestingAccountsSpendableBalan
 
 		alreadyVested := vAcc.EndTime < tmtime.Now().Unix()
 
-		balances, err := GetAccountSpendableBalance(s.cctx.GRPCClient, address)
+		balances, err := testfactory.GetAccountSpendableBalance(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 		expectedSpendableBal := initBalanceForGasFee
 		if alreadyVested {
@@ -140,7 +139,7 @@ func (s *VestingModuleTestSuite) TestGenesisDelayedVestingAccountsTransfer() {
 		name := s.unusedAccount(DelayedVestingAccountType)
 		address := getAddress(name, s.cctx.Keyring).String()
 
-		resAccBytes, err := GetRawAccountInfo(s.cctx.GRPCClient, address)
+		resAccBytes, err := testfactory.GetRawAccountInfo(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		var vAcc vestingtypes.DelayedVestingAccount
@@ -162,7 +161,7 @@ func (s *VestingModuleTestSuite) TestGenesisDelayedVestingAccountsDelegation() {
 		name := s.unusedAccount(DelayedVestingAccountType)
 		address := getAddress(name, s.cctx.Keyring).String()
 
-		resAccBytes, err := GetRawAccountInfo(s.cctx.GRPCClient, address)
+		resAccBytes, err := testfactory.GetRawAccountInfo(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		var vAcc vestingtypes.DelayedVestingAccount
@@ -186,7 +185,7 @@ func (s *VestingModuleTestSuite) TestGenesisDelayedVestingAccountsClaimDelegatio
 		name := s.unusedAccount(DelayedVestingAccountType)
 		address := getAddress(name, s.cctx.Keyring).String()
 
-		resAccBytes, err := GetRawAccountInfo(s.cctx.GRPCClient, address)
+		resAccBytes, err := testfactory.GetRawAccountInfo(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		var vAcc vestingtypes.DelayedVestingAccount
@@ -210,14 +209,14 @@ func (s *VestingModuleTestSuite) TestGenesisPeriodicVestingAccountsSpendableBala
 		name := s.unusedAccount(PeriodicVestingAccountType)
 		address := getAddress(name, s.cctx.Keyring).String()
 
-		resAccBytes, err := GetRawAccountInfo(s.cctx.GRPCClient, address)
+		resAccBytes, err := testfactory.GetRawAccountInfo(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		var vAcc vestingtypes.PeriodicVestingAccount
 		err = vAcc.Unmarshal(resAccBytes)
 		assert.NoError(s.T(), err)
 
-		balances, err := GetAccountSpendableBalance(s.cctx.GRPCClient, address)
+		balances, err := testfactory.GetAccountSpendableBalance(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		vestedCoins := vAcc.GetVestedCoins(tmtime.Now())
@@ -244,7 +243,7 @@ func (s *VestingModuleTestSuite) TestGenesisPeriodicVestingAccountsDelegation() 
 		name := s.unusedAccount(PeriodicVestingAccountType)
 		address := getAddress(name, s.cctx.Keyring).String()
 
-		resAccBytes, err := GetRawAccountInfo(s.cctx.GRPCClient, address)
+		resAccBytes, err := testfactory.GetRawAccountInfo(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		var vAcc vestingtypes.PeriodicVestingAccount
@@ -270,7 +269,7 @@ func (s *VestingModuleTestSuite) TestGenesisPeriodicVestingAccountsDelegationPar
 		name := s.unusedAccount(PeriodicVestingAccountType)
 		address := getAddress(name, s.cctx.Keyring).String()
 
-		resAccBytes, err := GetRawAccountInfo(s.cctx.GRPCClient, address)
+		resAccBytes, err := testfactory.GetRawAccountInfo(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		var vAcc vestingtypes.PeriodicVestingAccount
@@ -305,7 +304,7 @@ func (s *VestingModuleTestSuite) TestGenesisPeriodicVestingAccountsClaimDelegati
 		name := s.unusedAccount(PeriodicVestingAccountType)
 		address := getAddress(name, s.cctx.Keyring).String()
 
-		resAccBytes, err := GetRawAccountInfo(s.cctx.GRPCClient, address)
+		resAccBytes, err := testfactory.GetRawAccountInfo(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		var vAcc vestingtypes.PeriodicVestingAccount
@@ -329,14 +328,14 @@ func (s *VestingModuleTestSuite) TestGenesisContinuousVestingAccountsSpendableBa
 		name := s.unusedAccount(ContinuousVestingAccountType)
 		address := getAddress(name, s.cctx.Keyring).String()
 
-		resAccBytes, err := GetRawAccountInfo(s.cctx.GRPCClient, address)
+		resAccBytes, err := testfactory.GetRawAccountInfo(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		var vAcc vestingtypes.ContinuousVestingAccount
 		err = vAcc.Unmarshal(resAccBytes)
 		assert.NoError(s.T(), err)
 
-		balances, err := GetAccountSpendableBalance(s.cctx.GRPCClient, address)
+		balances, err := testfactory.GetAccountSpendableBalance(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		vestedCoins := vAcc.GetVestedCoins(tmtime.Now())
@@ -362,7 +361,7 @@ func (s *VestingModuleTestSuite) TestGenesisContinuousVestingAccountsDelegation(
 		name := s.unusedAccount(ContinuousVestingAccountType)
 		address := getAddress(name, s.cctx.Keyring).String()
 
-		resAccBytes, err := GetRawAccountInfo(s.cctx.GRPCClient, address)
+		resAccBytes, err := testfactory.GetRawAccountInfo(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		var vAcc vestingtypes.ContinuousVestingAccount
@@ -388,7 +387,7 @@ func (s *VestingModuleTestSuite) TestGenesisContinuousVestingAccountsDelegationP
 		name := s.unusedAccount(ContinuousVestingAccountType)
 		address := getAddress(name, s.cctx.Keyring).String()
 
-		resAccBytes, err := GetRawAccountInfo(s.cctx.GRPCClient, address)
+		resAccBytes, err := testfactory.GetRawAccountInfo(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		var vAcc vestingtypes.ContinuousVestingAccount
@@ -421,7 +420,7 @@ func (s *VestingModuleTestSuite) TestGenesisContinuousVestingAccountsClaimDelega
 		name := s.unusedAccount(ContinuousVestingAccountType)
 		address := getAddress(name, s.cctx.Keyring).String()
 
-		resAccBytes, err := GetRawAccountInfo(s.cctx.GRPCClient, address)
+		resAccBytes, err := testfactory.GetRawAccountInfo(s.cctx.GRPCClient, address)
 		assert.NoError(s.T(), err)
 
 		var vAcc vestingtypes.ContinuousVestingAccount
@@ -464,11 +463,11 @@ func (s *VestingModuleTestSuite) testTransferVestingAmount(name string) {
 func (s *VestingModuleTestSuite) testDelegatingVestingAmount(name string) {
 	address := getAddress(name, s.cctx.Keyring).String()
 
-	del, err := GetAccountDelegations(s.cctx.GRPCClient, address)
+	del, err := testfactory.GetAccountDelegations(s.cctx.GRPCClient, address)
 	assert.NoError(s.T(), err)
 	assert.Empty(s.T(), del, "initial delegation must be empty")
 
-	validators, err := GetValidators(s.cctx.GRPCClient)
+	validators, err := testfactory.GetValidators(s.cctx.GRPCClient)
 	assert.NoError(s.T(), err)
 	assert.NotEmpty(s.T(), validators)
 
@@ -482,10 +481,10 @@ func (s *VestingModuleTestSuite) testDelegatingVestingAmount(name string) {
 
 	resQ, err := s.cctx.WaitForTx(resTx.TxHash, 10)
 	assert.NoError(s.T(), err)
-	assert.EqualValues(s.T(), 0, resQ.TxResult.Code, "the delegation TX must succeed")
+	assert.EqualValues(s.T(), 0, resQ.TxResult.Code, fmt.Sprintf("the delegation TX must succeed: \n%s", resQ.TxResult.String()))
 
 	// verify the delegations
-	del, err = GetAccountDelegations(s.cctx.GRPCClient, address)
+	del, err = testfactory.GetAccountDelegations(s.cctx.GRPCClient, address)
 	assert.NoError(s.T(), err)
 	assert.NotEmpty(s.T(), del, "delegations must not be empty")
 	assert.EqualValues(s.T(),
@@ -516,14 +515,14 @@ func (s *VestingModuleTestSuite) testClaimDelegationReward(name string) {
 	rewardAmount := resR.Rewards[0].Reward.AmountOf(app.BondDenom).RoundInt().Int64()
 	assert.Greater(s.T(), rewardAmount, int64(0), "rewards must be more than zero")
 
-	balancesBefore, err := GetAccountSpendableBalance(s.cctx.GRPCClient, address)
+	balancesBefore, err := testfactory.GetAccountSpendableBalance(s.cctx.GRPCClient, address)
 	assert.NoError(s.T(), err)
 
 	// minExpectedBalance is used because more tokens may be vested to the
 	// account in the middle of this test
 	minExpectedBalance := balancesBefore.AmountOf(app.BondDenom).Int64() + rewardAmount
 
-	validators, err := GetValidators(s.cctx.GRPCClient)
+	validators, err := testfactory.GetValidators(s.cctx.GRPCClient)
 	assert.NoError(s.T(), err)
 	assert.NotEmpty(s.T(), validators, "empty validators set")
 
@@ -536,10 +535,10 @@ func (s *VestingModuleTestSuite) testClaimDelegationReward(name string) {
 
 	resQ, err := s.cctx.WaitForTx(resTx.TxHash, 10)
 	assert.NoError(s.T(), err)
-	assert.EqualValues(s.T(), 0, resQ.TxResult.Code, "the claim reward TX must succeed")
+	assert.EqualValues(s.T(), 0, resQ.TxResult.Code, fmt.Sprintf("the claim reward TX must succeed: \n%s", resQ.TxResult.String()))
 
 	// Check if the reward amount in the account
-	balancesAfter, err := GetAccountSpendableBalance(s.cctx.GRPCClient, address)
+	balancesAfter, err := testfactory.GetAccountSpendableBalance(s.cctx.GRPCClient, address)
 	assert.NoError(s.T(), err)
 
 	assert.GreaterOrEqual(s.T(), balancesAfter.AmountOf(app.BondDenom).Int64(), minExpectedBalance)
@@ -560,21 +559,15 @@ func (s *VestingModuleTestSuite) testClaimDelegationReward(name string) {
 func (s *VestingModuleTestSuite) initRegularAccounts(count int) testnode.GenesisOption {
 	names := testfactory.GenerateAccounts(count)
 	s.accounts.Store(RegularAccountType, accountDispenser{names: names})
-	bAccounts, balances := generateBaseAccounts(s.kr, names)
+	bAccounts, balances := testfactory.GenerateBaseAccounts(s.kr, names)
+	gAccounts := authtypes.GenesisAccounts{}
+	for i := range bAccounts {
+		gAccounts = append(gAccounts, &bAccounts[i])
+	}
 
 	return func(gs map[string]json.RawMessage) map[string]json.RawMessage {
-		gAccounts := authtypes.GenesisAccounts{}
-		for i := range bAccounts {
-			gAccounts = append(gAccounts, &bAccounts[i])
-		}
-
-		var err error
-		gs, err = AddAccountsToGenesisState(gs, gAccounts...)
+		gs, err := AddGenesisAccountsWithBalancesToGenesisState(gs, gAccounts, balances)
 		assert.NoError(s.T(), err)
-
-		gs, err = AddBalancesToGenesisState(gs, balances...)
-		assert.NoError(s.T(), err)
-
 		return gs
 	}
 }
@@ -595,33 +588,19 @@ func (s *VestingModuleTestSuite) initRegularAccounts(count int) testnode.Genesis
 // with varying end times using the generated accounts and balances, adds them to the genesis state,
 // and returns the modified genesis state.
 func (s *VestingModuleTestSuite) initDelayedVestingAccounts(count int) testnode.GenesisOption {
+	initCoinsForGasFee := sdk.NewCoin(app.BondDenom, sdk.NewInt(initBalanceForGasFee))
+
 	names := testfactory.GenerateAccounts(count)
 	s.accounts.Store(DelayedVestingAccountType, accountDispenser{names: names})
-	bAccounts, balances := generateBaseAccounts(s.kr, names)
+	vAccounts, balances := testfactory.GenerateDelayedVestingAccounts(s.kr, names, initCoinsForGasFee)
+	gAccounts := authtypes.GenesisAccounts{}
+	for i := range vAccounts {
+		gAccounts = append(gAccounts, vAccounts[i])
+	}
 
-	initCoinsForGasFee := sdk.NewCoin(app.BondDenom, sdk.NewInt(initBalanceForGasFee))
 	return func(gs map[string]json.RawMessage) map[string]json.RawMessage {
-		endTime := tmtime.Now().Add(-2 * time.Second)
-
-		gAccounts := authtypes.GenesisAccounts{}
-		for i := range bAccounts {
-			bal := findBalanceByAddress(balances, bAccounts[i].GetAddress().String())
-			assert.NotNil(s.T(), bal)
-			coins := bal.Coins.Sub(initCoinsForGasFee)
-			acc := vestingtypes.NewDelayedVestingAccount(&bAccounts[i], coins, endTime.Unix())
-			gAccounts = append(gAccounts, acc)
-
-			// the endTime is increased for each account to be able to test various scenarios
-			endTime = endTime.Add(5 * time.Second)
-		}
-
-		var err error
-		gs, err = AddAccountsToGenesisState(gs, gAccounts...)
+		gs, err := AddGenesisAccountsWithBalancesToGenesisState(gs, gAccounts, balances)
 		assert.NoError(s.T(), err)
-
-		gs, err = AddBalancesToGenesisState(gs, balances...)
-		assert.NoError(s.T(), err)
-
 		return gs
 	}
 }
@@ -640,41 +619,19 @@ func (s *VestingModuleTestSuite) initDelayedVestingAccounts(count int) testnode.
 // A GenesisOption function that takes the current genesis state as input, adds the generated accounts
 // and balances to it, and returns the modified genesis state.
 func (s *VestingModuleTestSuite) initPeriodicVestingAccounts(count int) testnode.GenesisOption {
+	initCoinsForGasFee := sdk.NewCoin(app.BondDenom, sdk.NewInt(initBalanceForGasFee))
+
 	names := testfactory.GenerateAccounts(count)
 	s.accounts.Store(PeriodicVestingAccountType, accountDispenser{names: names})
-	bAccounts, balances := generateBaseAccounts(s.kr, names)
-
-	allocationPerPeriod := int64(vestingAmount / 4)
-	periods := vestingtypes.Periods{
-		vestingtypes.Period{Length: int64(6), Amount: sdk.Coins{sdk.NewInt64Coin(app.BondDenom, allocationPerPeriod)}},
-		vestingtypes.Period{Length: int64(6), Amount: sdk.Coins{sdk.NewInt64Coin(app.BondDenom, allocationPerPeriod)}},
-		vestingtypes.Period{Length: int64(6), Amount: sdk.Coins{sdk.NewInt64Coin(app.BondDenom, allocationPerPeriod)}},
-		vestingtypes.Period{Length: int64(6), Amount: sdk.Coins{sdk.NewInt64Coin(app.BondDenom, allocationPerPeriod)}},
+	vAccounts, balances := testfactory.GeneratePeriodicVestingAccounts(s.kr, names, initCoinsForGasFee)
+	gAccounts := authtypes.GenesisAccounts{}
+	for i := range vAccounts {
+		gAccounts = append(gAccounts, vAccounts[i])
 	}
 
-	initCoinsForGasFee := sdk.NewCoin(app.BondDenom, sdk.NewInt(initBalanceForGasFee))
 	return func(gs map[string]json.RawMessage) map[string]json.RawMessage {
-		startTime := tmtime.Now()
-
-		gAccounts := authtypes.GenesisAccounts{}
-		for i := range bAccounts {
-			bal := findBalanceByAddress(balances, bAccounts[i].GetAddress().String())
-			assert.NotNil(s.T(), bal)
-			coins := bal.Coins.Sub(initCoinsForGasFee)
-			acc := vestingtypes.NewPeriodicVestingAccount(&bAccounts[i], coins, startTime.Unix(), periods)
-			gAccounts = append(gAccounts, acc)
-
-			// the startTime is increased for each account to be able to test various scenarios
-			startTime = startTime.Add(5 * time.Second)
-		}
-
-		var err error
-		gs, err = AddAccountsToGenesisState(gs, gAccounts...)
+		gs, err := AddGenesisAccountsWithBalancesToGenesisState(gs, gAccounts, balances)
 		assert.NoError(s.T(), err)
-
-		gs, err = AddBalancesToGenesisState(gs, balances...)
-		assert.NoError(s.T(), err)
-
 		return gs
 	}
 }
@@ -693,35 +650,19 @@ func (s *VestingModuleTestSuite) initPeriodicVestingAccounts(count int) testnode
 // A GenesisOption function that takes the current genesis state as input, adds the generated accounts
 // and balances to it, and returns the modified genesis state.
 func (s *VestingModuleTestSuite) initContinuousVestingAccounts(count int) testnode.GenesisOption {
+	initCoinsForGasFee := sdk.NewCoin(app.BondDenom, sdk.NewInt(initBalanceForGasFee))
 	names := testfactory.GenerateAccounts(count)
 	s.accounts.Store(ContinuousVestingAccountType, accountDispenser{names: names})
-	bAccounts, balances := generateBaseAccounts(s.kr, names)
 
-	initCoinsForGasFee := sdk.NewCoin(app.BondDenom, sdk.NewInt(initBalanceForGasFee))
+	vAccounts, balances := testfactory.GenerateContinuousVestingAccounts(s.kr, names, initCoinsForGasFee)
+	gAccounts := authtypes.GenesisAccounts{}
+	for i := range vAccounts {
+		gAccounts = append(gAccounts, vAccounts[i])
+	}
+
 	return func(gs map[string]json.RawMessage) map[string]json.RawMessage {
-		startTime := tmtime.Now()
-
-		gAccounts := authtypes.GenesisAccounts{}
-		for i := range bAccounts {
-			bal := findBalanceByAddress(balances, bAccounts[i].GetAddress().String())
-			assert.NotNil(s.T(), bal)
-			coins := bal.Coins.Sub(initCoinsForGasFee)
-
-			endTime := startTime.Add(10 * time.Second)
-			acc := vestingtypes.NewContinuousVestingAccount(&bAccounts[i], coins, startTime.Unix(), endTime.Unix())
-			gAccounts = append(gAccounts, acc)
-
-			// the startTime is increased for each account to be able to test various scenarios
-			startTime = startTime.Add(5 * time.Second)
-		}
-
-		var err error
-		gs, err = AddAccountsToGenesisState(gs, gAccounts...)
+		gs, err := AddGenesisAccountsWithBalancesToGenesisState(gs, gAccounts, balances)
 		assert.NoError(s.T(), err)
-
-		gs, err = AddBalancesToGenesisState(gs, balances...)
-		assert.NoError(s.T(), err)
-
 		return gs
 	}
 }
@@ -767,170 +708,34 @@ func (t accountType) String() string {
 	}
 }
 
-// AddAccountsToGenesisState adds the provided accounts to the genesis state (gs) map for the auth module.
-// It takes the raw genesis state (gs) and a variadic number of GenesisAccount objects (accounts) as inputs.
-// Then, it updates the given genesis state and returns it.
-func AddAccountsToGenesisState(gs map[string]json.RawMessage, accounts ...authtypes.GenesisAccount) (map[string]json.RawMessage, error) {
+// AddGenesisAccountsWithBalancesToGenesisState adds the given genesis accounts and balances to the
+// provided genesis state. It returns the updated genesis state and an error if any occurred.
+//
+// Parameters:
+// - gs: A map representing the current genesis state.
+// - gAccounts: A slice of genesis accounts to be added.
+// - balances: A slice of balances to be added.
+//
+// Returns:
+// - gs: The updated genesis state after adding the accounts and balances.
+// - error: An error if any occurred during the process.
+func AddGenesisAccountsWithBalancesToGenesisState(
+	gs map[string]json.RawMessage,
+	gAccounts []authtypes.GenesisAccount,
+	balances []banktypes.Balance,
+) (map[string]json.RawMessage, error) {
 	encCfg := encoding.MakeConfig(app.ModuleEncodingRegisters...)
-	var authGenState authtypes.GenesisState
-	err := encCfg.Codec.UnmarshalJSON(gs[authtypes.ModuleName], &authGenState)
+
+	var err error
+	gs, err = testfactory.AddAccountsToGenesisState(encCfg, gs, gAccounts...)
 	if err != nil {
 		return gs, err
 	}
 
-	pAccs, err := authtypes.PackAccounts(accounts)
+	gs, err = testfactory.AddBalancesToGenesisState(encCfg, gs, balances)
 	if err != nil {
 		return gs, err
 	}
-
-	// set the accounts in the genesis state
-	authGenState.Accounts = append(authGenState.Accounts, pAccs...)
-	gs[authtypes.ModuleName] = encCfg.Codec.MustMarshalJSON(&authGenState)
 
 	return gs, nil
-}
-
-// AddBalancesToGenesisState adds the provided balances to the genesis state (gs) for the bank module.
-// It takes the raw genesis state (gs) and a variadic number of Balance objects (balances) as inputs.
-// It returns the updated gs and nil if the process is successful.
-func AddBalancesToGenesisState(gs map[string]json.RawMessage, balances ...banktypes.Balance) (map[string]json.RawMessage, error) {
-	encCfg := encoding.MakeConfig(app.ModuleEncodingRegisters...)
-	var bankGenState banktypes.GenesisState
-	err := encCfg.Codec.UnmarshalJSON(gs[banktypes.ModuleName], &bankGenState)
-	if err != nil {
-		return gs, err
-	}
-
-	bankGenState.Balances = append(bankGenState.Balances, balances...)
-	gs[banktypes.ModuleName] = encCfg.Codec.MustMarshalJSON(&bankGenState)
-
-	return gs, nil
-}
-
-// GetValidators retrieves the validators from the staking module using the provided gRPC client connection (grpcConn).
-// It takes a gRPC client connection (grpcConn) as input.
-// Then, it returns the validators from the response.
-func GetValidators(grpcConn *grpc.ClientConn) (stakingtypes.Validators, error) {
-	scli := stakingtypes.NewQueryClient(grpcConn)
-	vres, err := scli.Validators(context.Background(), &stakingtypes.QueryValidatorsRequest{})
-
-	if vres == nil {
-		return stakingtypes.Validators{}, err
-	}
-	return vres.Validators, err
-}
-
-// GetAccountDelegations retrieves the delegations for the specified account address using the provided gRPC client connection (grpcConn).
-// It takes a gRPC client connection (grpcConn) and the account address (address) as inputs.
-// If an error occurs during the request, it returns nil and the error.
-// Otherwise, it returns the delegation responses from the response.
-func GetAccountDelegations(grpcConn *grpc.ClientConn, address string) (stakingtypes.DelegationResponses, error) {
-	cli := stakingtypes.NewQueryClient(grpcConn)
-	res, err := cli.DelegatorDelegations(context.Background(),
-		&stakingtypes.QueryDelegatorDelegationsRequest{DelegatorAddr: address})
-	if err != nil {
-		return nil, err
-	}
-
-	return res.DelegationResponses, err
-}
-
-// GetAccountSpendableBalance retrieves the spendable balance of an account for the specified address using gRPC.
-// It takes a gRPC client connection (grpcConn) and the account address (address) as inputs.
-// If the account is not found or an error occurs, it returns nil and the error.
-// Otherwise, it returns the spendable balances of the account as an sdk.Coins object.
-//
-// Parameters:
-// - grpcConn: A gRPC client connection.
-// - address: The account address to retrieve the spendable balances for.
-//
-// Returns:
-// The spendable balances of the account as an sdk.Coins object, or nil and an error if the retrieval fails.
-func GetAccountSpendableBalance(grpcConn *grpc.ClientConn, address string) (sdk.Coins, error) {
-	cli := banktypes.NewQueryClient(grpcConn)
-	res, err := cli.SpendableBalances(
-		context.Background(),
-		&banktypes.QuerySpendableBalancesRequest{
-			Address: address,
-		},
-	)
-	if err != nil || res == nil {
-		return nil, err
-	}
-	return res.GetBalances(), nil
-}
-
-// GetRawAccountInfo retrieves the raw account information for the specified address using gRPC.
-// It takes a gRPC client connection (grpcConn) and the account address (address) as inputs.
-// If no account found or an error occurs, it returns nil and the error.
-// Otherwise, it returns the value field of the account in the response, which represents the raw account information.
-//
-// Parameters:
-// - grpcConn: A gRPC client connection.
-// - address: The account address to retrieve the raw information for.
-//
-// Returns:
-// The raw account information as a byte slice, or nil and an error if the retrieval fails.
-func GetRawAccountInfo(grpcConn *grpc.ClientConn, address string) ([]byte, error) {
-	cli := authtypes.NewQueryClient(grpcConn)
-	res, err := cli.Account(context.Background(), &authtypes.QueryAccountRequest{
-		Address: address,
-	})
-
-	if err != nil || res == nil {
-		return nil, err
-	}
-
-	return res.Account.Value, nil
-}
-
-// generateBaseAccounts generates base accounts and their corresponding balances.
-// It takes a keyring.Keyring instance and a slice of account names as inputs.
-// For each name in the names slice, it creates a new base account and adds it to
-// the bAccounts slice. It also creates a banktypes.Balance struct for the account,
-// with the account's address and coins including an initial coin for gas fee.
-// The balances are added to the balances slice.
-//
-// Parameters:
-// - kr: A keyring.Keyring instance used to create base accounts.
-// - names: A slice of account names.
-//
-// Returns:
-// - bAccounts: A slice of authtypes.BaseAccount representing the generated base accounts.
-// - balances: A slice of banktypes.Balance representing the balances of the generated accounts.
-func generateBaseAccounts(kr keyring.Keyring, names []string) ([]authtypes.BaseAccount, []banktypes.Balance) {
-	initCoinsForGasFee := sdk.NewCoin(app.BondDenom, sdk.NewInt(initBalanceForGasFee))
-
-	bAccounts := []authtypes.BaseAccount{}
-	balances := []banktypes.Balance{}
-	for _, name := range names {
-		acc, coins := testfactory.NewBaseAccount(kr, name)
-		bAccounts = append(bAccounts, *acc)
-		balances = append(balances, banktypes.Balance{
-			Address: acc.GetAddress().String(),
-			Coins:   coins.Add(initCoinsForGasFee),
-		})
-	}
-
-	return bAccounts, balances
-}
-
-// findBalanceByAddress finds a balance in the given slice of banktypes.Balance
-// by matching the address with the provided address string.
-// It iterates over the balances slice and returns the pointer to the first balance
-// that has a matching address. If no matching balance is found, it returns nil.
-//
-// Parameters:
-// - balances: A slice of banktypes.Balance to search within.
-// - address: The address string to match against balance addresses.
-//
-// Returns:
-// A pointer to the banktypes.Balance with a matching address, or nil if not found.
-func findBalanceByAddress(balances []banktypes.Balance, address string) *banktypes.Balance {
-	for _, b := range balances {
-		if b.Address == address {
-			return &b
-		}
-	}
-	return nil
 }
